+++ src_3_0_0/app-store/actions.js
@@ -1,8 +1,7 @@
 import axios from "axios";
 import {initializeLayerList} from "@masterportal/masterportalapi/src/rawLayerList";
 import getNestedValues from "../utils/getNestedValues";
-import {getAndMergeRawLayersFilteredByMdId, getAndMergeRawLayer} from "./utils/getAndMergeRawLayer";
-import {modifyRawLayerList} from "./utils/modifyRawLayerList";
+import {getAndMergeAllRawLayers, getAndMergeRawLayer} from "./utils/getAndMergeRawLayer";
 
 export default {
     /**
@@ -39,6 +38,8 @@ export default {
 
     /**
      * Extends all layers of config.json with the attributes of the layer in services.json.
+     * If portalConfig.tree contains parameter 'layerIDsToIgnore', 'metaIDsToIgnore', 'metaIDsToMerge' or 'layerIDsToStyle' the raw layerlist is filtered and merged.
+     * Config entry portalConfig.tree.validLayerTypesAutoTree is respected.
      * Replaces the extended layer in state.layerConf.
      * @returns {void}
      */
@@ -46,7 +47,7 @@ export default {
         const layerContainer = getNestedValues(state.layerConfig, "Layer", "Ordner").flat(Infinity);
 
         if (state.portalConfig?.tree?.type === "auto") {
-            const rawLayers = getAndMergeRawLayersFilteredByMdId(state.portalConfig?.tree?.validLayerTypesAutoTree);
+            const rawLayers = getAndMergeAllRawLayers(state.portalConfig?.tree);
 
             commit("addToLayerConfig", {layerConfigs: {Layer: rawLayers}, parentKey: "Fachdaten"});
         }
@@ -76,23 +77,16 @@ export default {
 
     /**
      * Load the services.json via masterportalapi.
-     * If portalConfig.tree contains parameter 'layerIDsToIgnore', 'metaIDsToIgnore', 'metaIDsToMerge' or 'layerIDsToStyle' the raw layerlist is filtered.
      * @returns {void}
      */
     loadServicesJson ({state, commit}) {
-        initializeLayerList(state.configJs?.layerConf, (layerList, error) => {
+        initializeLayerList(state.configJs?.layerConf, (_, error) => {
             if (error) {
                 // Implementieren, wenn Alert da ist:
                 // Radio.trigger("Alert", "alert", {
                 //     text: "<strong>Die Datei '" + layerConfUrl + "' konnte nicht geladen werden!</strong>",
                 //     kategorie: "alert-warning"
                 // });
-            }
-            else if (state.portalConfig?.tree) {
-                const rawLayerArray = modifyRawLayerList(layerList, state.portalConfig?.tree);
-
-                initializeLayerList(rawLayerArray,
-                    () => commit("setLoadedConfigs", "servicesJson"));
             }
             else {
                 commit("setLoadedConfigs", "servicesJson");