+++ src_3_0_0/app-store/utils/getAndMergeRawLayer.js
@@ -1,4 +1,4 @@
-import {getLayerWhere, getLayerList} from "@masterportal/masterportalapi/src/rawLayerList";
+import {getLayerWhere, getLayerList, initializeLayerList} from "@masterportal/masterportalapi/src/rawLayerList";
 import omit from "../../utils/omit";
 /**
  * Returns the extended raw layer to the id contained in the layer configuration.
@@ -24,7 +24,7 @@ function mergeRawLayer (layerConf, rawLayer) {
 
     if (layerConf) {
         if (Array.isArray(layerConf.id)) {
-            mergedLayer = mergeLayerByIds(layerConf);
+            mergedLayer = mergeLayerWithSeveralIds(layerConf);
         }
         else if (layerConf.children) {
             mergedLayer = fillGroupLayer(layerConf);
@@ -76,19 +76,18 @@ function fillGroupLayer (layerConf) {
 /**
  * Merges layer configuration with ids property of type array.
  * @param {Object} layerConf configuartion of layer like in the config.json with ids in an array
- * @returns {Object} the merged raw layer
+ * @returns {Object|undefined} the merged raw layer or undefined if layer cannot be merged
  */
-function mergeLayerByIds (layerConf) {
+function mergeLayerWithSeveralIds (layerConf) {
     // refactored from parser.js, mergeexistingLayers and parserCustomTree.js, parseTree
     const ids = layerConf.id,
         existingLayers = [],
         maxScales = [],
-        minScales = [],
-        layerlist = getLayerList();
+        minScales = [];
     let mergedLayer = {};
 
     ids?.forEach(id => {
-        const layer = layerlist.find(aLayer => aLayer.id === id);
+        const layer = getLayerWhere({id: id});
 
         if (layer) {
             existingLayers.push(layer);
@@ -115,26 +114,52 @@ function mergeLayerByIds (layerConf) {
 }
 
 /**
- * Returns all layer to add to states layerConfig 'Fachdaten' for treetype 'auto'.
+ * Returns all layer to add to states layerConfig for treetype 'auto'.
  * Filters the raw layerlist by typ and datasets.
  * Creates new raw layer if datasets contains more than one entry.
- * @param {Array} validLayerTypes layer types to be used with the tree.type 'auto'. If not configured,  ["WMS", "SENSORTHINGS", "TERRAIN3D", "TILESET3D", "OBLIQUE"] are used
+ * Filters and merges layerList by config parameters from portalConfig.tree 'layerIDsToIgnore', 'metaIDsToIgnore', 'metaIDsToMerge' and 'layerIDsToStyle'.
+ * Layers with ids contained in 'layerIDsToIgnore'are removed from list.
+ * Layers with meta_ids contained in 'metaIDsToIgnore'are removed from list.
+ * Layers with meta_ids contained in 'metaIDsToMerge'are are merged to one layer.
+ * Properties in 'layerIDsToStyle'are are assigned to dedicated layers. If an entry has more than one style, new layers are created for each style.
+ * Ids are composed by origin id and styles value.
+ * @param  {Object[]} treeConfig - portalConfig.tree from state, may contain 'layerIDsToIgnore', 'metaIDsToIgnore', 'metaIDsToMerge' and 'layerIDsToStyle',
+ * If config 'validLayerTypes' is set, it contains layer types to be used with the tree.type 'auto'. If not configured,  ["WMS", "SENSORTHINGS", "TERRAIN3D", "TILESET3D", "OBLIQUE"] are used.
  * @returns {Array} the filtered layer configurations
  */
-export function getAndMergeRawLayersFilteredByMdId (validLayerTypes) {
-    // refactored from parserDefaultTree.js
+export function getAndMergeAllRawLayers (treeConfig) {
+    // refactored from parserDefaultTree.js and layerList.js
     const layerList = getLayerList(),
         rawLayers = [],
-        layerTypes = validLayerTypes ? validLayerTypes : ["WMS", "SENSORTHINGS", "TERRAIN3D", "TILESET3D", "OBLIQUE"];
+        validLayerTypes = treeConfig?.validLayerTypesAutoTree ? treeConfig?.validLayerTypesAutoTree : ["WMS", "SENSORTHINGS", "TERRAIN3D", "TILESET3D", "OBLIQUE"],
+        layerIDsToIgnore = Object.prototype.hasOwnProperty.call(treeConfig, "layerIDsToIgnore") ? treeConfig.layerIDsToIgnore : [],
+        metaIDsToIgnore = Object.prototype.hasOwnProperty.call(treeConfig, "metaIDsToIgnore") ? treeConfig.metaIDsToIgnore : [],
+        metaIDsToMerge = Object.prototype.hasOwnProperty.call(treeConfig, "metaIDsToMerge") ? treeConfig.metaIDsToMerge : [],
+        layerIDsToStyle = Object.prototype.hasOwnProperty.call(treeConfig, "layerIDsToStyle") ? treeConfig.layerIDsToStyle : [],
+        idsOfLayersToStyle = layerIDsToStyle.map(entry => entry.id),
+        toMergeByMdId = {};
     let relatedWMSLayerIds = [];
 
     for (let i = 0; i < layerList.length; i++) {
-        const rawLayer = layerList[i];
+        const rawLayer = layerList[i],
+            layerType = rawLayer.typ?.toUpperCase();
 
-        if (!layerTypes.includes(rawLayer.typ?.toUpperCase()) || !rawLayer.datasets || rawLayer.datasets.length === 0) {
+        if (!validLayerTypes.includes(layerType) ||
+            !rawLayer.datasets ||
+            rawLayer.datasets.length === 0 ||
+            layerIDsToIgnore.includes(rawLayer.id) ||
+            metaIDsToIgnore.includes(rawLayer.datasets[0].md_id)) {
             continue;
         }
-        if (rawLayer.typ.toUpperCase() === "SENSORTHINGS" && rawLayer.related_wms_layers !== undefined) {
+        if (layerType === "WMS" && metaIDsToMerge.includes(rawLayer.datasets[0].md_id)) {
+            collectLayersToMergeByMetaId(toMergeByMdId, rawLayer.datasets[0].md_id, rawLayer);
+            continue;
+        }
+        if (idsOfLayersToStyle.includes(rawLayer.id)) {
+            styleAndMergeLayers(layerIDsToStyle, rawLayer, rawLayers);
+            continue;
+        }
+        if (layerType === "SENSORTHINGS" && rawLayer.related_wms_layers !== undefined) {
             relatedWMSLayerIds = relatedWMSLayerIds.concat(rawLayer.related_wms_layers);
         }
         if (rawLayer.datasets.length > 1) {
@@ -150,7 +175,10 @@ export function getAndMergeRawLayersFilteredByMdId (validLayerTypes) {
             rawLayers.push(rawLayer);
         }
     }
+    mergeByMetaIds(toMergeByMdId, rawLayers);
     removeFromLayerList(relatedWMSLayerIds, rawLayers);
+    // set rawLayerList in masterportalapi to this rawLayers
+    initializeLayerList(rawLayers);
     return rawLayers;
 }
 
@@ -167,5 +195,83 @@ function removeFromLayerList (ids = [], layerList = []) {
         if (index > -1) {
             layerList.splice(index, 1);
         }
+    });
+}
+/**
+ * Properties in 'layerIDsToStyle'are are assigned to dedicated layers. If an entry has more than one style, new layers are created for each style.
+ * Ids are composed by origin id and styles value.
+ * @param  {Array} layerIDsToStyle contains style configurations
+ * @param  {Object} layer raw layer from services.json
+ * @param  {Object[]} [layerList=[]] The layers from services.json
+ * @returns {void}
+ */
+function styleAndMergeLayers (layerIDsToStyle, layer, layerList) {
+    const styleConfig = layerIDsToStyle.find(item => item.id === layer.id),
+        rawLayer = Object.assign(layer, styleConfig);
+
+    if (rawLayer.typ === "WMS" && Array.isArray(rawLayer.styles) && rawLayer.styles.length > 1) {
+        rawLayer.styles.forEach(function (style, index) {
+            const cloneObj = Object.assign({}, rawLayer);
+
+            cloneObj.style = style;
+            cloneObj.legendURL = rawLayer.legendURL[index];
+            cloneObj.name = rawLayer.name[index];
+            cloneObj.id = rawLayer.id + rawLayer.styles[index];
+            cloneObj.styles = rawLayer.styles[index];
+            layerList.push(cloneObj);
+        });
+    }
+    else {
+        layerList.push(rawLayer);
+    }
+}
+/**
+ * Collects layers to merge by metaId.
+ * @param {Object} toMergeByMdId key contains metaId and value contains an Array of layers with same metaId
+ * @param  {String} mdId metaId of the layers first dataset
+ * @param {Object} rawLayer raw layer from services.json
+ * @returns {void}
+ */
+function collectLayersToMergeByMetaId (toMergeByMdId, mdId, rawLayer) {
+    if (!Array.isArray(toMergeByMdId[mdId])) {
+        toMergeByMdId[mdId] = [];
+    }
+    toMergeByMdId[mdId].push(rawLayer);
+}
+
+/**
+ * Adds configured attributes from the config.js parameter 'layerIDsToStyle' to the objects via the Id.
+ * From layers with multiple styles, a new layer is created per style and added to layerList.
+ * @param {Object} toMergeByMdId key contains metaId and value contains an Array of layers with same metaId
+ * @param  {Object[]} [layerList=[]] The layers from services.json
+ * @returns {void}
+ */
+function mergeByMetaIds (toMergeByMdId, layerList) {
+    Object.values(toMergeByMdId).forEach(layers => {
+        const layersContent = [];
+        let gfiAttributesSet = false,
+            maxScale,
+            minScale,
+            mergedLayer = null;
+
+        layers.forEach((layer, index) => {
+            if (index === 0) {
+                mergedLayer = Object.assign({}, layer);
+                mergedLayer.name = layer.datasets[0].md_name;
+                maxScale = layer.maxScale;
+                minScale = layer.minScale;
+            }
+            if (layer.gfiAttributes !== "ignore" && !gfiAttributesSet) {
+                mergedLayer.gfiAttributes = layer.gfiAttributes;
+                gfiAttributesSet = true;
+            }
+            layersContent.push(layer.layers);
+            maxScale = Math.max(maxScale, layer.maxScale);
+            minScale = Math.min(minScale, layer.minScale);
+        });
+        mergedLayer.layers = layersContent.join(",");
+        mergedLayer.maxScale = maxScale;
+        mergedLayer.minScale = minScale;
+        layerList.push(mergedLayer);
     });
 }